util-docker: Update how rocBLAS is built in gcn-gpu Dockerfile
rocBLAS builds for all targets by default, but it requires we set the
HCC_AMDGPU_TARGET env variable to all the targets we want to build for.

This change explicitly sets all the gfx versions we intend to support in
gem5, and removes the extraneous "-a all" from the install command.

Additionally, doing this allowed for DNNMark to run on gfx902.

Change-Id: Id1a00433beaa23f6935c12073d5bddc38431886b
Reviewed-on: https://gem5-review.googlesource.com/c/public/gem5/+/54623
Reviewed-by: Matt Sinclair <mattdsinclair@gmail.com>
Maintainer: Matt Sinclair <mattdsinclair@gmail.com>
Reviewed-by: Matthew Poremba <matthew.poremba@amd.com>
Tested-by: kokoro <noreply+kokoro@google.com>